+++ filter.go
@@ -54,6 +54,7 @@ type Filter struct {
 	To time.Time
 
 	// IncludeTime sets whether the selected period should contain the time (hour, minute, second).
+	// The Analyzer will ignore this and just always use the date without time, except where it makes sense.
 	IncludeTime bool
 
 	// Period sets the period to group results.